+++ openedge-plugin/src/test/java/org/sonar/plugins/openedge/sensor/InvalidXMLFilterStreamTest.java
@@ -10,7 +10,6 @@ import java.util.Set;
 import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
 
-import org.sonar.plugins.openedge.sensor.OpenEdgeXREFSensor.InvalidXMLFilterStream;
 import org.testng.annotations.Test;
 import org.xml.sax.SAXException;
 import org.xml.sax.SAXParseException;
